+++ detekt/wrapper/src/main/kotlin/bazel/detekt/Platform.kt
@@ -1,0 +1,13 @@
+package bazel.detekt
+
+import kotlin.system.exitProcess
+
+internal interface Platform {
+
+    fun exit(code: Int)
+
+    class Impl : Platform {
+
+        override fun exit(code: Int) = exitProcess(code)
+    }
+}
