+++ detekt/wrapper/src/main/kotlin/bazel/detekt/Result.kt
@@ -1,0 +1,11 @@
+package bazel.detekt
+
+internal sealed class Result {
+    object Success : Result()
+    data class Failure(val description: String) : Result()
+
+    val consoleStatusCode = when (this) {
+        is Success -> 0
+        is Failure -> 1
+    }
+}
