+++ detekt/wrapper/src/main/kotlin/bazel/detekt/main.kt
@@ -1,0 +1,25 @@
+@file:JvmName("Main")
+
+package bazel.detekt
+
+import io.reactivex.schedulers.Schedulers
+
+/**
+ * The wrapper purpose:
+ *
+ * - switch between the continuous worker mode and the single-shot regular mode;
+ * - pass arguments to Detekt;
+ * - silence Detekt output on successful executions.
+ */
+fun main(arguments: Array<String>) {
+    val executable = Executable.DetektImpl(Detekt.Impl())
+    val consoleStreams = Streams.system()
+
+    val application = if ("--persistent_worker" in arguments) {
+        Application.Worker(Schedulers.io(), WorkerExecutable.Impl(executable), WorkerStreams.Impl(consoleStreams))
+    } else {
+        Application.OneShot(executable, consoleStreams, Platform.Impl())
+    }
+
+    application.run(arguments)
+}
