--- detekt/wrapper/src/main/kotlin/io/buildfoundation/bazel/rulesdetekt/wrapper/Detekt.kt
@@ -1,16 +1,0 @@
-package io.buildfoundation.bazel.rulesdetekt.wrapper
-
-import io.gitlab.arturbosch.detekt.cli.buildRunner
-import java.io.PrintStream
-
-internal interface Detekt {
-
-    fun execute(args: Array<String>, outputPrinter: PrintStream, errorPrinter: PrintStream)
-
-    class Impl : Detekt {
-
-        override fun execute(args: Array<String>, outputPrinter: PrintStream, errorPrinter: PrintStream) {
-            buildRunner(args, outputPrinter, errorPrinter).execute()
-        }
-    }
-}
