--- detekt/wrapper/src/main/kotlin/io/buildfoundation/bazel/rulesdetekt/wrapper/Platform.kt
@@ -1,13 +1,0 @@
-package io.buildfoundation.bazel.rulesdetekt.wrapper
-
-import kotlin.system.exitProcess
-
-internal interface Platform {
-
-    fun exit(code: Int)
-
-    class Impl : Platform {
-
-        override fun exit(code: Int) = exitProcess(code)
-    }
-}
