--- detekt/wrapper/src/main/kotlin/io/buildfoundation/bazel/rulesdetekt/wrapper/WorkerExecutable.kt
@@ -1,27 +1,0 @@
-package io.buildfoundation.bazel.rulesdetekt.wrapper
-
-import bazel.worker.WorkerProtocol.WorkRequest
-import bazel.worker.WorkerProtocol.WorkResponse
-
-internal interface WorkerExecutable {
-
-    fun execute(request: WorkRequest): WorkResponse
-
-    class Impl(private val executable: Executable) : WorkerExecutable {
-
-        override fun execute(request: WorkRequest): WorkResponse {
-            val result = executable.execute(request.argumentsList.toTypedArray())
-
-            val resultOutput = when (result) {
-                is Result.Success -> WorkResponse.getDefaultInstance().output
-                is Result.Failure -> result.description
-            }
-
-            return WorkResponse.newBuilder()
-                    .setRequestId(request.requestId)
-                    .setOutput(resultOutput)
-                    .setExitCode(result.consoleStatusCode)
-                    .build()
-        }
-    }
-}
