+++ detekt/wrapper/src/test/kotlin/bazel/detekt/TestsSuite.kt
@@ -1,0 +1,13 @@
+package bazel.detekt
+
+import org.junit.runner.RunWith
+import org.junit.runners.Suite
+
+@RunWith(Suite::class)
+@Suite.SuiteClasses(
+        DetektExecutableTests::class,
+        OneShotApplicationTests::class,
+        WorkerExecutableTests::class,
+        WorkerStreamsTests::class
+)
+class TestsSuite
