+++ detekt/wrapper/src/testFixtures/kotlin/bazel/detekt/TestDetekt.kt
@@ -1,0 +1,16 @@
+package bazel.detekt
+
+import java.io.PrintStream
+
+internal class TestDetekt(private val executionResult: ExecutionResult) : Detekt {
+
+    enum class ExecutionResult {
+        Success, Failure
+    }
+
+    override fun execute(args: Array<String>, outputPrinter: PrintStream, errorPrinter: PrintStream) {
+        if (executionResult == ExecutionResult.Failure) {
+            throw RuntimeException()
+        }
+    }
+}
