+++ detekt/wrapper/src/testFixtures/kotlin/bazel/detekt/TestPlatform.kt
@@ -1,0 +1,12 @@
+package bazel.detekt
+
+import bazel.detekt.Platform
+
+internal class TestPlatform : Platform {
+
+    var exitCode = Int.MIN_VALUE
+
+    override fun exit(code: Int) {
+        exitCode = code
+    }
+}
