--- detekt/wrapper/src/testFixtures/kotlin/io/buildfoundation/bazel/rulesdetekt/wrapper/TestDetekt.kt
@@ -1,16 +1,0 @@
-package io.buildfoundation.bazel.rulesdetekt.wrapper
-
-import java.io.PrintStream
-
-internal class TestDetekt(private val executionResult: ExecutionResult) : Detekt {
-
-    enum class ExecutionResult {
-        Success, Failure
-    }
-
-    override fun execute(args: Array<String>, outputPrinter: PrintStream, errorPrinter: PrintStream) {
-        if (executionResult == ExecutionResult.Failure) {
-            throw RuntimeException()
-        }
-    }
-}
