--- detekt/wrapper/src/testFixtures/kotlin/io/buildfoundation/bazel/rulesdetekt/wrapper/TestExecutable.kt
@@ -1,6 +1,0 @@
-package io.buildfoundation.bazel.rulesdetekt.wrapper
-
-internal class TestExecutable(private val result: Result) : Executable {
-
-    override fun execute(args: Array<String>) = result
-}
