--- detekt/wrapper/src/testFixtures/kotlin/io/buildfoundation/bazel/rulesdetekt/wrapper/TestPlatform.kt
@@ -1,10 +1,0 @@
-package io.buildfoundation.bazel.rulesdetekt.wrapper
-
-internal class TestPlatform : Platform {
-
-    var exitCode = Int.MIN_VALUE
-
-    override fun exit(code: Int) {
-        exitCode = code
-    }
-}
