+++ tests/integration/src/main/kotlin/bazel/detekt/B.kt
@@ -1,0 +1,3 @@
+package bazel.detekt
+
+class B
