+++ tests/integration/src/main/kotlin/bazel/detekt/C.kt
@@ -1,0 +1,3 @@
+package bazel.detekt
+
+class C
