+++ tests/integration/src/main/kotlin/bazel/detekt/main.kt
@@ -1,0 +1,5 @@
+package bazel.detekt
+
+fun main(args: Array<String>) {
+
+}
