--- tests/integration/src/main/kotlin/io/buildfoundation/bazelrulesdetekt/integrationtest/A.kt
@@ -1,3 +1,0 @@
-package io.buildfoundation.bazelrulesdetekt.integrationtest
-
-class A
