--- tests/integration/src/main/kotlin/io/buildfoundation/bazelrulesdetekt/integrationtest/Main.kt
@@ -1,5 +1,0 @@
-package io.buildfoundation.bazelrulesdetekt.integrationtest
-
-fun main(args: Array<String>) {
-
-}
